requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
requests==2.32.0
pydantic==2.9.0
pyjwt==2.12.1
bcrypt==5.0.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.11
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
